Conexion a la base de datos funcionando

+++ src/main/java/com/ecommerce/api/dto/CartItemResponse.java
@@ -2,13 +2,15 @@ package com.ecommerce.api.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class CartItemResponse {
     Long id;
     Long productId;
     String productName;
     String imageUrl;
-    Double price;
+    BigDecimal price;
     Integer quantity;
-    Double subtotal;
+    BigDecimal subtotal;
 }
